Working support for various types of Identity values (BIGINT, INT, SMALLINT, TINYINT, Numeric(?, 0)), etc. Not all scenarios have been tested, but the implementation along with some performance improvements are now implemented and all regression Integration Tests are passing.

+++ NetStandard.SqlBulkHelpers/SqlBulkHelper/QueryProcessing/SqlBulkHelpersDataReader.cs
@@ -14,7 +14,7 @@ namespace SqlBulkHelpers
         
         private IEnumerator<T> _dataEnumerator;
         private Dictionary<string, int> _processingDefinitionOrdinalDictionary;
-        private long _entityCounter = 0;
+        private int _entityCounter = 0;
 
         public SqlBulkHelpersDataReader(IEnumerable<T> entityData, SqlBulkHelpersProcessingDefinition processingDefinition, SqlBulkHelpersTableDefinition tableDefinition)
         {
